+++ tests/core/pyspec/eth2spec/gen_helpers/gen_base/typing.py
